+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
-import App from './containers/App';
+import App from './components/App';
 import rootReducer from './reducers/index';
 
 const obtainRandomId = () => Math.floor(Math.random() * 1000);
